requirements.txt
amqp==2.3.2
arrow==0.12.1
billiard==3.5.0.4
binaryornot==0.4.4
cachetools==2.1.0
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-filters==0.2.1
django-heroku==0.3.1
djangorestframework==3.8.2
flashtext==2.7
future==0.16.0
gin==0.1.6
gunicorn==19.9.0
idna==2.7
Jinja2==2.10
jinja2-time==0.2.0
kombu==4.2.1
MarkupSafe==1.0
numpy==1.15.0
opencv-python==3.4.2.17
Pillow==5.2.0
poyo==0.4.1
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
scipy==1.1.0
six==1.11.0
urllib3==1.23
vine==1.1.4
virtualenv==16.0.0
Wave==0.0.2
whichcraft==0.4.1
whitenoise==4.1
